% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{hash2json}
\alias{hash2json}
\title{Hash to JSON converter}
\usage{
hash2json(hash)
}
\arguments{
\item{hash}{a (possibly nested) named list}
}
\value{
a data.frame that behaves as expected with \code{jsonlite::toJSON}.
}
\description{
Turns a (possibly nested) named list into a \code{\link{data.frame}} ready to
be converted into JSON by \code{\link{toJSON}}.
}
