% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multicolumn}
\alias{multicolumn}
\title{Display content in multiple columns automatically}
\usage{
multicolumn(ncols, ..., unit_element = NULL)
}
\arguments{
\item{ncols}{number of columns}

\item{...}{content to spread}

\item{unit_element}{a CSS selector (e.g. a tag name or a class) that
specifies the elements that you want to spread over columns. This is used
try to not break them accross multiple columns but rather keep them
together in the same one.}
}
\description{
This function uses CSS to spread content accross multiple columns. It
provides a more automatic solution as opposed to manually defining columns
using Shiny's \code{\link[shiny]{column}} function.
}
