% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-loader.R
\name{tabular_data_loader}
\alias{tabular_data_loader}
\title{Tabular data loader}
\usage{
tabular_data_loader(input, output, session, id)
}
\arguments{
\item{input}{the shiny \code{input} object}

\item{output}{the shiny \code{output} object}

\item{session}{the shiny \code{session} object}

\item{id}{}
}
\value{
a \code{\link{data.frame}} wrapped inside a reactive context
}
\description{
Tabular data loader
}
\seealso{
\code{\link{tabular_data_loader_input}} for the input function
}
