% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-loader.R
\name{tabular_data_loader_input}
\alias{tabular_data_loader_input}
\title{Tabular data loader input}
\usage{
tabular_data_loader_input(
  id,
  file_label = "Select file",
  allowed_field_separators = c(TAB = "\\t", Space = " ", Comma = ",", Semicolon = ";"),
  allowed_field_delimiters = c(`Double quote` = "\\"", `Single quote` = "'", None = ""),
  allow_header_toggle = FALSE,
  allow_rownames_toggle = FALSE,
  na_string = "NA",
  allowed_decimal_separators = c(Dot = ".", Comma = ","),
  allow_transposition = FALSE
)
}
\arguments{
\item{id}{an identifier unique in the namespace}

\item{file_label}{the label for the file input}

\item{allowed_field_separators}{the characters that may be used to separate
fields within a row}

\item{allowed_field_delimiters}{the characters that may be used for field
quoting (see the parameter \code{quote} in \code{\link{read.table}})}

\item{allow_header_toggle}{whether to give the option to treat the first row
as a header}

\item{allow_rownames_toggle}{wether to give the option to treat the first
column as a header}

\item{allowed_decimal_separators}{the characters that may be used to separate
the integral and fractional parts in a number}

\item{allow_transposition}{whether to give the option to transpose the table}
}
\value{
a list of tags that can be included in a shiny UI definition
}
\description{
A set of inputs that allows loading tabular data into an R
\code{\link{data.frame}}.
}
\details{
For any of the parameters, \code{NULL} may be passed and in that case the
particular control will by hidden from the user and the default value in
\code{\link{read.table}} will be used. For boolean parameters like
\code{allow_transposition} a non-\code{NULL} value will be used as the
default for that control. For the rest of parameters, a list of options must
be provided per the instructions in the \code{choices} parameter description
in \code{link{radioButtons}} or the like. In a list of choices, the first one
is the default value.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

  ui <- fluidPage(
    tabular_data_loader_input("tabular"),
    tableOutput("tabular_output")
  )
  server <- function(input, output) {
    df <- callModule(tabular_data_loader, "tabular")
    output$tabular_output <- renderTable(df(), rownames = TRUE)
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link{tabular_data_loader}} for the server function
}
