% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{ti}
\alias{ti}
\title{Table input (server function)}
\usage{
ti(input, output, session, id, fields, data)
}
\arguments{
\item{input}{the shiny \code{input} object}

\item{output}{the shiny \code{output} object}

\item{session}{the shiny \code{session} object}

\item{id}{the id given to the input}

\item{fields}{a list of fields (see \code{\link{ti_inputs}}) to include on
each row}

\item{data}{a matrix representing the state of the fields}
}
\description{
Table input (server function)
}
\seealso{
\code{\link{ti_input}} for the UI function and
  \code{\link{update_ti}} for updating the ti on the fly (similarly to
  \code{updateSelectInput} for instance).
}
