% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{ti_assets}
\alias{ti_assets}
\title{Table input asset manifest}
\usage{
ti_assets()
}
\description{
An \code{\link{htmlDependency}} object describing the assets (JS and CSS
files) needed for this module to run.
}
