% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{ti_input}
\alias{ti_input}
\title{Tabular input}
\usage{
ti_input(id)
}
\arguments{
\item{id}{the id to give to this instance of a table input}
}
\value{
a list of tags that can be used as part of a shiny UI definition
}
\description{
A table which has inputs on its cells
}
\seealso{
\code{\link{ti}} for the server part of this module
}
