% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{ti_inputs}
\alias{ti_inputs}
\alias{ti_label}
\alias{ti_checkbox}
\alias{ti_dropdown}
\alias{ti_radio}
\title{Inputs for table input (ti)}
\usage{
ti_label(name)

ti_checkbox(name)

ti_dropdown(name, options)

ti_radio(name, options)
}
\arguments{
\item{name}{the input name}

\item{options}{a list of options for radio and dropdown fields matching the
format of the parameter \code{choices} in \code{\link{radioButtons}}.}
}
\description{
Inputs for table input (ti)
}
