% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-input.R
\name{update_ti}
\alias{update_ti}
\title{Update a table input}
\usage{
update_ti(session, id = "namespace", fields, data)
}
\arguments{
\item{session}{the shiny session object}

\item{id}{the id of the table input to update}

\item{fields}{a list of the fields present in each row}

\item{data}{a matrix representing the state of the fields}
}
\description{
Update a table input
}
\seealso{
\code{\link{ti}}
}
